+++ src/veterinaryClinic/core/drugStore/components/Penicilin.java
@@ -3,7 +3,7 @@ package veterinaryClinic.core.drugStore.components;
 import veterinaryClinic.core.drugStore.Component;
 
 public class Penicilin extends Component {
-    public Penicilin(String title, Double weight, int power) {
+    public Penicilin(String title, Double weight, Double power) {
         super(title, weight, power);
     }
 }
