+++ src/veterinaryClinic/core/drugStore/components/Water.java
@@ -3,7 +3,7 @@ package veterinaryClinic.core.drugStore.components;
 import veterinaryClinic.core.drugStore.Component;
 
 public class Water extends Component {
-    public Water(String title, Double weight, int power) {
+    public Water(String title, Double weight, Double power) {
         super(title, weight, power);
     }
 }
